refactor: Move simulateRevert from InstanceAgentsInternal to GameInstanceInternal
Commit 400 :D
