+++ src/agents/instance-agents-internal.ts
@@ -114,15 +114,6 @@ export interface InstanceAgentsInternal {
      * instance reverting. Make sure to use this correctly.
      */
     scrubAgents(): void;
-
-    /**
-     * Builds and executes a `TrackedEvent` that reverts all changes in this
-     * `InstanceAgentsInternal` since a specified event.
-     *
-     * @param source The agent history on which the revert function will be based. Is not modified.
-     * @param revertTo The id of the `TrackedEvent` to which the state will be reverted.
-     */
-    simulateRevert(source: InstanceAgentsInternal, revertTo?: number): void;
 }
 
 /** Whether the property is a positive integer, meaning its a valid agent id. */
